Fixed font size issues on Mac OS X (hopefully)
I manually removed any font-size styles in the about box. Unfortunately,
Qt Creator automatically puts them in, so this might be an ongoing pain
in the ass to deal with. Hopefully I won't ever have to change anything
on that screen. Anyway, as long as the sizes are left out, Qt defaults
to using whatever the system font is, which is the appropriate font for
that screen.