This is an incomplete implementation of the new "erase some of the SIMM" feature. The erase capability is there, so next up comes the write.
The next step is to add a mechanism to the protocol to start writing at
a location other than 0 (although this is not needed for this particular
feature, I'd like to have it in there first). Then, I have to update
this software to use that new protocol and make sure verification still
works. I'm about 50% done I'd say.